Made changes to structure and added seller button
Changed the sellerPages folder to be set to be inside of websitePages. Changed the buyer page and buyer.js files to include a "Want to Sell?" button that redirects to an application page where a wannabe seller can apply for the seller role through emailing an admin ( this functionality is still a WIP).