added seller refund view and admin transaction overview
seller can now see when a buyer requests a refund. it's a little janky currently but it does work, gracie is the one who did it. I did the seller notifications as we were both crashing out on different things. she also added a way to see the recent transactions as an admin, and I did some small tweaks with this functionality.