Adding Pagination to Main Page
I added pagination to the main page rather than having a set number of items show up. It allows a user to click on a next or previous button or click on a specific page number. I also went in and fixed how products show up on the page using search.